+++ internal/providers/pluginfw/common/common.go
@@ -1,10 +1,8 @@
 package common
 
 import (
-	"context"
 	"fmt"
 
-	"github.com/databricks/databricks-sdk-go/useragent"
 	"github.com/databricks/terraform-provider-databricks/common"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
@@ -56,12 +54,4 @@ func GetDatabricksStagingName(name string) string {
 
 func GetDatabricksProductionName(name string) string {
 	return fmt.Sprintf("databricks_%s", name)
-}
-
-func SetResourceNameInContext(ctx context.Context, resourceName string) context.Context {
-	return useragent.InContext(ctx, "resource", resourceName)
-}
-
-func SetDataSourceNameInContext(ctx context.Context, dataSourceName string) context.Context {
-	return useragent.InContext(ctx, "data", dataSourceName)
 }
