+++ internal/providers/pluginfw/common/common_test.go
@@ -1,10 +1,8 @@
 package common
 
 import (
-	"context"
 	"testing"
 
-	"github.com/databricks/databricks-sdk-go/useragent"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -20,22 +18,4 @@ func TestGetDatabricksProductionName(t *testing.T) {
 	expected := "databricks_test"
 	result := GetDatabricksProductionName(resourceName)
 	assert.Equal(t, expected, result, "GetDatabricksProductionName should return the expected production name")
-}
-
-func TestSetResourceNameInContext(t *testing.T) {
-	ctx := context.Background()
-	resourceKey := "resource"
-	resourceName := "test-resource"
-	actualContext := SetResourceNameInContext(ctx, resourceName)
-	expectedContext := useragent.InContext(ctx, resourceKey, resourceName)
-	assert.Equal(t, expectedContext, actualContext)
-}
-
-func TestSetDataSourceNameInContext(t *testing.T) {
-	ctx := context.Background()
-	dataSourceKey := "data"
-	dataSourceName := "test-datasource"
-	actualContext := SetDataSourceNameInContext(ctx, dataSourceName)
-	expectedContext := useragent.InContext(ctx, dataSourceKey, dataSourceName)
-	assert.Equal(t, expectedContext, actualContext)
 }
